actually its better to just have all of this together
